// ArchiCore/src/app/api-client/model/userResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleResponse } from './roleResponse';


export interface UserResponse { 
    id?: number;
    username?: string;
    name?: string;
    surname?: string;
    phone?: string;
    status?: string;
    createdAt?: string;
    role?: RoleResponse;
}